all: remove mod.rs and nested mods
