--- src/agent_data/mod.rs
@@ -1,2 +1,0 @@
-pub mod agent_data;
-pub use agent_data::AgentData;
