--- src/configuration_wrapper/mod.rs
@@ -1,2 +1,0 @@
-pub mod configuration_wrapper;
-pub use configuration_wrapper::ConfigurationWrapper;
