--- src/http_server/mod.rs
@@ -1,3 +1,0 @@
-pub mod http_server;
-pub use http_server::MyFilesState;
-pub mod routes;
