--- src/http_server/routes/mod.rs
@@ -1,6 +1,0 @@
-pub mod get_files;
-pub mod get_status;
-pub mod hello_world;
-pub use get_files::get_files;
-pub use get_status::get_status;
-pub use hello_world::hello_world;
