--- src/lister/mod.rs
@@ -1,2 +1,0 @@
-pub mod lister;
-pub use lister::list_directories;
