--- src/logger/mod.rs
@@ -1,2 +1,0 @@
-pub mod logger;
-pub use logger::init_logger;
