--- src/my_files/mod.rs
@@ -1,3 +1,0 @@
-pub mod my_files;
-pub use my_files::MyFiles;
-pub use my_files::MyFilesBuilder;
