--- src/watcher/mod.rs
@@ -1,2 +1,0 @@
-pub mod watcher;
-pub use watcher::watch_directories;
